giam sat api

+++ src/main/java/org/example/hdmm/repository/CoQuanThueRepository.java
@@ -3,5 +3,5 @@ package org.example.hdmm.repository;
 import org.example.hdmm.models.CoQuanThue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoQuanThueRepository extends JpaRepository<CoQuanThue, Integer> {
+public interface CoQuanThueRepository extends JpaRepository<CoQuanThue, Long> {
 }
