+++ src/main/java/org/example/hdmm/repository/KyQuayThuongRepository.java
@@ -10,19 +10,14 @@ import java.util.List;
 
 public interface KyQuayThuongRepository extends JpaRepository<KyQuayThuong, Long> {
     @Query("SELECT CASE WHEN COUNT (e) >0 THEN true ELSE false END FROM KyQuayThuong e WHERE e.maKy = :maKy AND e.coQuanThue.cqt = :cqt")
-    boolean existsByMaKyAndCoQuanThue_Id(@Param("maKy") Integer maKy,@Param("cqt") Integer cqt);
+    boolean existsByMaKyAndCoQuanThue_Id(@Param("maKy") Integer maKy,@Param("cqt") Long cqt);
     @Query("SELECT CASE WHEN COUNT(e) > 0 THEN true ELSE false END FROM KyQuayThuong e " +
             "WHERE e.tuNgay <= :date AND e.denNgay >= :date AND e.coQuanThue.cqt = :cqt ")
-    boolean existsByDate(@Param("date") Date date , @Param("cqt") Integer cqt);
-
-
+    boolean existsByDate(@Param("date") Date date , @Param("cqt") Long cqt);
     @Query("SELECT CASE WHEN COUNT(e) >0 THEN true ELSE false END FROM KyQuayThuong e " +
             "WHERE e.tuNgay <= :date AND e.denNgay >= :date AND e.coQuanThue.cqt = :cqt AND e.id != :id ")
     boolean exists1ByDate(@Param("date") Date date , @Param("cqt") Long cqt,@Param("id")Long id);
-
-
-
-
     @Query("select e from KyQuayThuong e where e.coQuanThue.cqt = :cqt")
     List<KyQuayThuong> findAllByCQT(@Param("cqt") Integer cqt);
+
 }
